Fix bug with losing focus on input field
